Split ImportService into pipeline-stage files
Pure relocation of import/service.rs into an import/service/ directory: the
command-loop and watcher stay in mod.rs, and the pipeline-stage method clusters
move into sibling files (format_prep, cover_image, reconcile, progress) plus the
inline tests. Methods move verbatim; only cross-module reach gets a pub(super)
bump, and the now-redundant per-item platform cfg is dropped in favor of gating
the module declarations.
