+++ bae-core/src/import/service/mod.rs
@@ -9,21 +9,14 @@ use crate::library::LibraryManager;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 use {
     crate::db::{
-        DbAlbum, DbAlbumArtist, DbFile, DbImport, DbRelease, DbReleaseMetadata, DbTrack,
-        DbTrackArtist,
+        DbAlbum, DbAlbumArtist, DbFile, DbRelease, DbReleaseMetadata, DbTrack, DbTrackArtist,
     },
     crate::import::folder_registry::ImportFolderRegistry,
     crate::import::folder_scanner::{
         scan_for_candidates_with_callback, InvalidCandidate, ScanItem,
     },
-    crate::import::handle::{fetch_artist_images, remap_album_artists, remap_track_artists},
     crate::import::track_to_file_mapper::map_tracks_to_files,
-    crate::import::types::{
-        CoverSelection, CueAudioAnalysis, CueFlacAnalysis, DiscoveredFile, ImportPhase,
-        PrepareStep, TrackFile,
-    },
-    crate::util::content_type::ContentType,
-    crate::util::content_type_hint::ContentTypeHint,
+    crate::import::types::{CoverSelection, DiscoveredFile, ImportPhase, PrepareStep, TrackFile},
     notify::RecursiveMode,
     notify_debouncer_full::{new_debouncer, DebounceEventResult},
     std::collections::{HashMap, HashSet},
@@ -36,6 +29,18 @@ use {
 use tokio::sync::{broadcast, mpsc};
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 use tracing::{debug, error, info, warn};
+
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
+mod cover_image;
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
+mod format_prep;
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
+mod progress;
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
+mod reconcile;
+
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
+use format_prep::resolve_file_content_type;
 
 /// Metadata-prep output for a folder import.
 ///
@@ -60,199 +65,12 @@ struct PreparedMetadata {
     artist_name: String,
 }
 
-/// Resolve the probe-verified `ContentType` for a discovered file.
-///
-/// Audio files are probed (only the decoder knows which codec the container
-/// holds — `.m4a` could be ALAC or AAC, and the extension alone can't tell).
-/// Non-audio files get their `ContentType` derived from the extension hint:
-/// that's an honest mapping for images (an extension on image bytes predicts
-/// the codec), text, and PDF. Anything the hint can't classify becomes
-/// `OctetStream`, which flows through the DB as-is.
-#[cfg(not(any(target_os = "ios", target_os = "android")))]
-fn resolve_file_content_type(path: &Path) -> Result<ContentType, String> {
-    let ext = path
-        .extension()
-        .and_then(|e| e.to_str())
-        .ok_or_else(|| format!("File has no extension: {:?}", path))?;
-    let hint = ContentTypeHint::from_extension(ext);
-
-    if hint.is_audio() {
-        let path_str = path
-            .to_str()
-            .ok_or_else(|| format!("Invalid path: {}", path.display()))?;
-        let probe = crate::audio_codec::probe_audio_from_path(path_str)
-            .ok_or_else(|| format!("Failed to probe audio file: {}", path.display()))?;
-        return Ok(probe.content_type);
-    }
-
-    Ok(match hint {
-        ContentTypeHint::Jpeg => ContentType::Jpeg,
-        ContentTypeHint::Png => ContentType::Png,
-        ContentTypeHint::Gif => ContentType::Gif,
-        ContentTypeHint::Webp => ContentType::Webp,
-        ContentTypeHint::Bmp => ContentType::Bmp,
-        ContentTypeHint::Svg => ContentType::Svg,
-        ContentTypeHint::PlainText => ContentType::PlainText,
-        ContentTypeHint::Pdf => ContentType::Pdf,
-        // Audio hints were handled above; anything else is unknown binary.
-        _ => ContentType::OctetStream,
-    })
-}
-
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 fn storage_mode_label(mode: &StorageMode) -> &'static str {
     match mode {
         StorageMode::Remote => "remote",
         StorageMode::Local => "local",
     }
-}
-
-/// Build an audio format for a track inside a CUE-backed file (shared FLAC or APE image).
-#[cfg(not(any(target_os = "ios", target_os = "android")))]
-fn cue_backed_audio_format(
-    db_track_id: &str,
-    file_path: &Path,
-    cue_pair: &CueFlacAnalysis,
-    cue_index: usize,
-    id: String,
-    now: chrono::DateTime<chrono::Utc>,
-) -> Result<crate::db::DbAudioFormat, String> {
-    use crate::db::DbAudioFormat;
-
-    let cue_track = cue_pair.cue_sheet.tracks.get(cue_index).ok_or_else(|| {
-        format!(
-            "CUE track index {} out of bounds for {}",
-            cue_index,
-            file_path.display()
-        )
-    })?;
-
-    let pregap_ms = cue_track
-        .pregap_duration_ms()
-        .filter(|&ms| ms > 0)
-        .map(|ms| ms as i64);
-
-    // Every CUE codec decodes its shared file natively and is trimmed to the
-    // track's sample window -- one shape across FLAC, APE and ALAC.
-    let fmt = cue_analysis_format(cue_pair);
-    let start_sample = cue_track.audio_start_sample(fmt.sample_rate);
-    let end_sample = cue_track.end_sample(fmt.sample_rate);
-    Ok(DbAudioFormat::new(
-        db_track_id,
-        fmt.content_type,
-        fmt.sample_rate as i64,
-        fmt.bits_per_sample,
-        fmt.channels as i64,
-        start_sample as i64,
-        end_sample.map(|s| s as i64),
-        id,
-        now,
-    )
-    .with_pregap(pregap_ms))
-}
-
-/// Build an audio format for a per-track file (FLAC, MP3, APE, etc.) via FFmpeg probe.
-#[cfg(not(any(target_os = "ios", target_os = "android")))]
-fn standalone_probed_audio_format(
-    db_track_id: &str,
-    file_path: &Path,
-    id: String,
-    now: chrono::DateTime<chrono::Utc>,
-) -> Result<crate::db::DbAudioFormat, String> {
-    use crate::db::DbAudioFormat;
-
-    let path_str = file_path
-        .to_str()
-        .ok_or_else(|| format!("Invalid path: {}", file_path.display()))?;
-    let probe = crate::audio_codec::probe_audio_from_path(path_str)
-        .ok_or_else(|| format!("Failed to probe audio file: {}", file_path.display()))?;
-
-    // A per-track file is its own whole-file window: (0, None) samples and the
-    // default (0, None) byte span -- the whole file.
-    Ok(DbAudioFormat::new(
-        db_track_id,
-        probe.content_type,
-        probe.sample_rate as i64,
-        probe.bits_per_sample.map(|b| b as i64),
-        probe.channels as i64,
-        0,
-        None,
-        id,
-        now,
-    ))
-}
-
-/// The audio format descriptor of a CUE-backed file, read once from its analysis
-/// and shared by every caller that needs any of these fields, so the per-codec
-/// extraction lives in one match instead of being repeated per field.
-#[cfg(not(any(target_os = "ios", target_os = "android")))]
-struct CueAnalysisFormat {
-    content_type: ContentType,
-    sample_rate: u32,
-    bits_per_sample: Option<i64>,
-    channels: u32,
-}
-
-#[cfg(not(any(target_os = "ios", target_os = "android")))]
-fn cue_analysis_format(cue_pair: &CueFlacAnalysis) -> CueAnalysisFormat {
-    match &cue_pair.analysis {
-        CueAudioAnalysis::Flac { flac_info } => CueAnalysisFormat {
-            content_type: ContentType::Flac,
-            sample_rate: flac_info.sample_rate,
-            bits_per_sample: Some(flac_info.bits_per_sample as i64),
-            channels: flac_info.channels,
-        },
-        CueAudioAnalysis::Ape { ape_info } => CueAnalysisFormat {
-            content_type: ContentType::Ape,
-            sample_rate: ape_info.sample_rate,
-            bits_per_sample: Some(ape_info.bits_per_sample as i64),
-            channels: ape_info.channels as u32,
-        },
-        CueAudioAnalysis::Alac {
-            sample_rate,
-            channels,
-            bits_per_sample,
-            ..
-        } => CueAnalysisFormat {
-            content_type: ContentType::Alac,
-            sample_rate: *sample_rate,
-            bits_per_sample: bits_per_sample.map(|b| b as i64),
-            channels: *channels,
-        },
-    }
-}
-
-/// Byte offset of each CUE track's end within the shared file, found by seeking
-/// (no decode) -- computed once per file. `ends[i]` is track `i`'s end byte; the
-/// last track has no entry and runs to EOF. The boundary is each track's
-/// `end_sample` -- the same sample `cue_backed_audio_format` trims to -- so the
-/// read-ahead ceiling and the decode window can't drift. `None` when the offsets
-/// can't be read (non-UTF-8 path, a non-last track missing its end sample, or a
-/// failed seek), distinct from a real empty result; the caller then spans the
-/// whole file.
-#[cfg(not(any(target_os = "ios", target_os = "android")))]
-fn cue_track_byte_ends(file_path: &Path, cue_pair: &CueFlacAnalysis) -> Option<Vec<u64>> {
-    let sample_rate = cue_analysis_format(cue_pair).sample_rate;
-    let tracks = &cue_pair.cue_sheet.tracks;
-    // Seek to each non-last track's end sample (the last track runs to EOF, so it
-    // has no end sample); every other track must have one.
-    let end_samples = match tracks
-        .iter()
-        .take(tracks.len().saturating_sub(1))
-        .map(|t| t.end_sample(sample_rate))
-        .collect::<Option<Vec<u64>>>()
-    {
-        Some(samples) => samples,
-        None => {
-            warn!("cue_track_byte_ends: a non-last track has no end sample in {file_path:?}");
-            return None;
-        }
-    };
-    let Some(path) = file_path.to_str() else {
-        warn!("cue_track_byte_ends: non-UTF-8 path, cannot seek byte offsets: {file_path:?}");
-        return None;
-    };
-    crate::audio_codec::frame_byte_offsets(path, &end_samples)
 }
 
 /// Send an import event on the broadcast bus, logging on send failure.
@@ -740,183 +558,6 @@ impl ImportService {
                 },
             );
         }
-    }
-
-    /// Reconcile the prepared release against existing library state,
-    /// record the import, and remap parsed artist IDs to their actual
-    /// DB IDs.
-    ///
-    /// Input is the already-mapped `ParsedAlbum` plus the raw metadata
-    /// pairs (empty for Unknown). The caller chooses the mapper based
-    /// on identity choice — `prepare_release` for Exact / Approximate,
-    /// `map_file_tags_to_db` for Unknown. This function does pure DB
-    /// work and string remapping, no network.
-    ///
-    /// Applies the user's `identity_choice` post-process on top of the
-    /// mapper output:
-    ///
-    /// - **Exact** — mapper output as-is. Identity rows keep their
-    ///   `source_release_id`; pressing-level metadata (year, format,
-    ///   label, catalog_number, country) seeds from the picked
-    ///   release.
-    /// - **Approximate** — identity rows get `source_release_id = None`
-    ///   (cross-source rows from url-rels mirror the user's choice and
-    ///   also become group-only); pressing-level metadata is cleared
-    ///   so the release row reflects "user didn't claim a specific
-    ///   pressing." Album-group-stable fields (title, artist, tracks)
-    ///   stay populated from the picked release.
-    /// - **Unknown** — mapper output (empty identity vec, file-tag
-    ///   release fields) passes through. `metadata_source` is
-    ///   `'file_tags'`; `metadata_source_release_id` stays NULL.
-    ///   The album lookup is skipped (empty identities), so the
-    ///   release lands on a fresh album.
-    ///
-    /// For Exact / Approximate, `metadata_source` and
-    /// `metadata_source_release_id` are kept pointing at the picked
-    /// release — the release records which source release seeded it
-    /// regardless of identity claim. For Unknown those columns
-    /// already arrive set by `map_file_tags_to_db`.
-    ///
-    /// `user_edit` is an optional overlay from the confirmation-page
-    /// editor. Applied after the choice transformation so the user's
-    /// edits win over the seeded values.
-    ///
-    /// Shared between folder and CD imports; callers handle the parts
-    /// that differ (cover art, file discovery, ripping).
-    async fn reconcile_prepared_release(
-        &self,
-        parsed: crate::import::ParsedAlbum,
-        resolved_metadata: Vec<(String, String)>,
-        import_id: &str,
-        source_path: &str,
-        identity_choice: &crate::import::IdentityChoice,
-        user_edit: Option<crate::import::ReleaseUserEdit>,
-    ) -> Result<PreparedMetadata, String> {
-        let library_manager = &self.library_manager;
-
-        let crate::import::ParsedAlbum {
-            album: mut db_album,
-            release: mut db_release,
-            tracks: mut db_tracks,
-            mut artists,
-            mut album_artists,
-            mut track_artists,
-            identities: parsed_identities,
-        } = parsed;
-
-        // Apply the identity choice on top of the mapper's output.
-        // Exact / Unknown preserve the mapper's per-source release IDs
-        // (Unknown's vec is empty anyway); Approximate NULLs them and
-        // clears the pressing-level cluster so the seeded record
-        // reflects "user claimed an album, not a pressing."
-        let identities = apply_identity_choice(&parsed_identities, identity_choice);
-        if matches!(
-            identity_choice,
-            crate::import::IdentityChoice::Approximate { .. }
-        ) {
-            // `disc_id` is a signal-domain field, not part of the
-            // pressing cluster — the import pipeline supplies it from
-            // the user's actual LOG/CUE artifacts separately, so wiping
-            // the source's value keeps the seeded record's signals
-            // reflecting only the user's physical media.
-            db_release.pressing = crate::db::Pressing::blank();
-            db_release.disc_id = None;
-        }
-
-        // User-edit overlay: apply the user's edits on top of the seed.
-        // Done here so reseeded fields (Approximate cleared) can still
-        // be user-overridden via the editor before commit. Returns
-        // updated db_album / db_release / db_tracks / album_artists /
-        // track_artists / artists with new artist rows merged in.
-        if let Some(edit) = user_edit {
-            apply_user_edit_to_seed(
-                &edit,
-                &mut db_album,
-                &mut db_release,
-                &mut db_tracks,
-                &mut artists,
-                &mut album_artists,
-                &mut track_artists,
-                library_manager.clock().as_ref(),
-                library_manager.ids().as_ref(),
-            )?;
-        }
-
-        let album_title = db_album.title.clone();
-        let artist_name = artists
-            .iter()
-            .find(|a| a.id == db_album.artist_id)
-            .expect("primary artist must be in artists vec")
-            .name
-            .clone();
-
-        let existing_album_id = library_manager
-            .find_existing_album_for_import(&identities)
-            .await?;
-        if let Some(album_id) = &existing_album_id {
-            db_release.album_id = album_id.clone();
-        }
-
-        let db_import = DbImport::new(
-            import_id,
-            &album_title,
-            &artist_name,
-            source_path,
-            library_manager.clock().now(),
-        );
-        library_manager
-            .insert_import(&db_import)
-            .await
-            .map_err(|e| format!("Failed to create import record: {}", e))?;
-
-        let resolved = library_manager
-            .find_or_create_artists(&artists)
-            .await
-            .map_err(|e| format!("Failed to resolve artists: {e}"))?;
-
-        let artist_id_map: HashMap<String, String> = artists
-            .iter()
-            .zip(resolved.iter())
-            .map(|(a, id)| (a.id.clone(), id.clone()))
-            .collect();
-
-        let remapped_primary_artist_id = artist_id_map
-            .get(&db_album.artist_id)
-            .ok_or_else(|| {
-                format!(
-                    "Primary artist ID {} not found in artist map",
-                    db_album.artist_id
-                )
-            })?
-            .clone();
-        db_album.artist_id = remapped_primary_artist_id;
-
-        let remapped_track_artists = remap_track_artists(&track_artists, &artist_id_map)?;
-        let remapped_album_artists = if existing_album_id.is_none() {
-            remap_album_artists(&album_artists, &artist_id_map)?
-        } else {
-            vec![]
-        };
-
-        let discogs_client = library_manager
-            .discogs_client()
-            .map_err(|e| format!("Failed to read Discogs key: {e}"))?;
-        if let Some(ref discogs_client) = discogs_client {
-            fetch_artist_images(library_manager, discogs_client, &artists, &artist_id_map).await;
-        }
-
-        Ok(PreparedMetadata {
-            db_album,
-            db_release,
-            db_tracks,
-            resolved_metadata,
-            existing_album_id,
-            remapped_track_artists,
-            remapped_album_artists,
-            identities,
-            album_title,
-            artist_name,
-        })
     }
 
     /// Prepare and run a folder import.
@@ -1575,255 +1216,6 @@ impl ImportService {
         Ok(())
     }
 
-    fn emit_started(&self, candidate_key: &str, release_id: &str, import_id: &str) {
-        send_event(
-            &self.event_tx,
-            crate::import::handle::ImportEvent::ImportProgress {
-                candidate_key: candidate_key.to_string(),
-                progress: ImportProgress::Started {
-                    id: release_id.to_string(),
-                    import_id: Some(import_id.to_string()),
-                },
-            },
-        );
-    }
-
-    /// Emit a coarse running-phase progress event for the candidate row. `id` is
-    /// the release id (or a track id, during the per-file referencing pass);
-    /// `percent` fills the candidate's determinate bar for that phase.
-    fn emit_phase_progress(
-        &self,
-        candidate_key: &str,
-        id: &str,
-        percent: u8,
-        phase: ImportPhase,
-        import_id: &str,
-    ) {
-        send_event(
-            &self.event_tx,
-            crate::import::handle::ImportEvent::ImportProgress {
-                candidate_key: candidate_key.to_string(),
-                progress: ImportProgress::Progress {
-                    id: id.to_string(),
-                    percent,
-                    phase,
-                    import_id: Some(import_id.to_string()),
-                },
-            },
-        );
-    }
-
-    /// Emit a loudness-measurement tick for the candidate's confirm pane. Routed
-    /// to a native leaf view (not the coarse candidate row), so the sub-track
-    /// cadence never churns the row. `fraction` is overall scan progress (0..1)
-    /// for the determinate bar; `tracks_done`/`tracks_total` label which track.
-    #[cfg(any(target_os = "macos", target_os = "linux", target_os = "windows"))]
-    fn emit_loudness_progress(
-        &self,
-        candidate_key: &str,
-        tracks_done: u32,
-        tracks_total: u32,
-        fraction: f32,
-    ) {
-        send_event(
-            &self.event_tx,
-            crate::import::handle::ImportEvent::ImportLoudnessProgress {
-                candidate_key: candidate_key.to_string(),
-                tracks_done,
-                tracks_total,
-                fraction,
-            },
-        );
-    }
-
-    /// Build a cover image record from local files without writing to DB.
-    /// Read the chosen cover file's bytes and build its `DbLibraryImage` record,
-    /// returning `(record, bytes)`. The caller hands the bytes to coven's local
-    /// store (the cover's home as a host-provided Local blob) and the record to
-    /// finalize; nothing is written to a bae path here.
-    #[allow(clippy::type_complexity)]
-    fn build_cover_image_record(
-        &self,
-        release_id: &str,
-        discovered_files: &[DiscoveredFile],
-        cover_image_path: Option<&Path>,
-    ) -> Result<Option<(crate::db::DbLibraryImage, Vec<u8>)>, String> {
-        use crate::db::{DbLibraryImage, LibraryImageType};
-
-        let image_extensions = ["jpg", "jpeg", "png", "gif", "webp"];
-        let mut image_files: Vec<(&DiscoveredFile, String)> = Vec::new();
-        for f in discovered_files {
-            let is_image = f
-                .path
-                .extension()
-                .and_then(|e| e.to_str())
-                .map(|e| image_extensions.contains(&e.to_lowercase().as_str()))
-                .unwrap_or(false);
-            if is_image {
-                let relative_path = self.get_relative_image_path(&f.path)?;
-                image_files.push((f, relative_path));
-            }
-        }
-
-        if image_files.is_empty() {
-            return Ok(None);
-        }
-
-        // Determine which file is the cover: match by absolute path if provided
-        let cover_index = if let Some(selected_path) = cover_image_path {
-            let found = image_files
-                .iter()
-                .position(|(f, _)| f.path.as_path() == selected_path);
-            if found.is_none() {
-                info!(
-                    "Selected cover {:?} not found among images, using priority",
-                    selected_path
-                );
-            }
-            found
-        } else {
-            None
-        };
-
-        let cover_index = cover_index.unwrap_or_else(|| {
-            image_files.sort_by(|(_, a), (_, b)| {
-                let a_priority = Self::image_cover_priority(a);
-                let b_priority = Self::image_cover_priority(b);
-                a_priority.cmp(&b_priority)
-            });
-            0
-        });
-
-        let (cover_file, relative_path) = &image_files[cover_index];
-        let content_type = resolve_file_content_type(&cover_file.path)?;
-        let source_url = format!("release://{}", relative_path);
-
-        // Read the cover bytes from the user's folder; the caller stores them in
-        // coven's local store and writes the row.
-        let bytes = match std::fs::read(&cover_file.path) {
-            Ok(bytes) => bytes,
-            Err(e) => {
-                error!(
-                    "Failed to read cover art {}: {e}",
-                    cover_file.path.display()
-                );
-                return Ok(None);
-            }
-        };
-
-        let now = self.library_manager.clock().now();
-        let db_image = DbLibraryImage {
-            id: release_id.to_string(),
-            image_type: LibraryImageType::Cover,
-            content_type,
-            file_size: bytes.len() as i64,
-            width: None,
-            height: None,
-            source: "local".to_string(),
-            source_url: Some(source_url),
-            // Computed in the finalize transaction under a browsable home.
-            cloud_path: None,
-            created_at: now,
-        };
-
-        Ok(Some((db_image, bytes)))
-    }
-
-    fn image_cover_priority(filename: &str) -> u8 {
-        let lower = filename.to_lowercase();
-        if lower.contains("cover") || lower.contains("front") {
-            return 0;
-        }
-        1
-    }
-
-    fn get_relative_image_path(&self, path: &std::path::Path) -> Result<String, String> {
-        let filename = path
-            .file_name()
-            .and_then(|n| n.to_str())
-            .ok_or_else(|| format!("Invalid filename: {:?}", path))?;
-        if let Some(parent) = path.parent() {
-            if let Some(parent_name) = parent.file_name().and_then(|n| n.to_str()) {
-                if parent_name == "scans" || parent_name == "artwork" || parent_name == "images" {
-                    return Ok(format!("{}/{}", parent_name, filename));
-                }
-            }
-        }
-        Ok(filename.to_string())
-    }
-
-    /// Build audio format records for all tracks. CUE-backed tracks already hold
-    /// their shared analysis + index; standalone tracks are probed here.
-    fn build_audio_formats(
-        tracks_to_files: &[TrackFile],
-        file_ids: &HashMap<PathBuf, String>,
-        clock: &dyn coven::Clock,
-        ids: &dyn coven::IdProvider,
-    ) -> Result<Vec<crate::db::DbAudioFormat>, String> {
-        let now = clock.now();
-        let mut audio_formats = Vec::with_capacity(tracks_to_files.len());
-        // Track end byte offsets per shared CUE file, computed once and reused
-        // across that file's tracks (one ffmpeg open per file, not per track).
-        // `None` means the offsets couldn't be read for that file.
-        let mut cue_ends_by_file: HashMap<PathBuf, Option<Vec<u64>>> = HashMap::new();
-
-        for track_file in tracks_to_files {
-            // Each track carries the absolute path to its source file; that path
-            // is the `file_ids` key — no bare-filename lookup that could collide
-            // across disc subfolders.
-            let file_id = file_ids.get(track_file.file_path()).ok_or_else(|| {
-                format!(
-                    "No DbFile registered for track source {:?}",
-                    track_file.file_path()
-                )
-            })?;
-
-            let format = match track_file {
-                TrackFile::CueBacked {
-                    db_track,
-                    file_path,
-                    cue_pair,
-                    cue_index,
-                } => {
-                    let af = cue_backed_audio_format(
-                        &db_track.id,
-                        file_path,
-                        cue_pair,
-                        *cue_index,
-                        ids.new_id(),
-                        now,
-                    )?;
-                    // The last track (no entry) runs to EOF, as does any track
-                    // whose offsets couldn't be read -- both keep the default
-                    // whole-file span.
-                    let ends = cue_ends_by_file.entry(file_path.clone()).or_insert_with(|| {
-                        let computed = cue_track_byte_ends(file_path, cue_pair);
-                        if computed.is_none() {
-                            warn!(
-                                "track byte offsets unavailable for {:?}; its tracks keep the whole-file read-ahead span",
-                                file_path
-                            );
-                        }
-                        computed
-                    });
-                    let end_byte = ends
-                        .as_ref()
-                        .and_then(|e| e.get(*cue_index))
-                        .map(|&b| b as i64);
-                    af.with_end_byte(end_byte)
-                }
-                TrackFile::Standalone {
-                    db_track,
-                    file_path,
-                } => standalone_probed_audio_format(&db_track.id, file_path, ids.new_id(), now)?,
-            };
-
-            audio_formats.push(format.with_file_id(file_id));
-        }
-
-        Ok(audio_formats)
-    }
-
     /// Measure each track's loudness + true peak and the album's combined
     /// loudness, attaching the per-track measurements to `audio_formats` and
     /// returning the album-level `(loudness_lufs, peak_linear)`.
@@ -2258,546 +1650,4 @@ pub(crate) async fn prepare_release(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use coven::FixedClock;
-    use coven::SequentialIdProvider;
-
-    #[test]
-    fn affected_roots_maps_changed_paths_to_their_watched_roots() {
-        let root_a = PathBuf::from("/music/new rips");
-        let root_b = PathBuf::from("/downloads/bandcamp");
-        let roots = vec![root_a.clone(), root_b.clone()];
-
-        // A change inside one root flags only that root.
-        let changed = [Path::new("/music/new rips/Album/01.flac")];
-        assert_eq!(affected_roots(&changed, &roots), vec![root_a.clone()]);
-
-        // Changes under both roots flag both, in roots order, deduped.
-        let changed = [
-            Path::new("/downloads/bandcamp/X/cover.jpg"),
-            Path::new("/music/new rips/Y"),
-            Path::new("/music/new rips/Z"),
-        ];
-        assert_eq!(affected_roots(&changed, &roots), vec![root_a, root_b]);
-
-        // A change outside every watched root flags nothing.
-        let changed = [Path::new("/elsewhere/file")];
-        assert!(affected_roots(&changed, &roots).is_empty());
-    }
-
-    /// `common_ancestor` derives the local-path root by folding over the
-    /// files' parent dirs. It must compare path components, not string
-    /// prefixes, so `/m/Album` and `/m/Album2` collapse to `/m` (a string
-    /// prefix would wrongly keep `/m/Album`), and an ancestor argument returns
-    /// itself rather than descending.
-    #[test]
-    fn common_ancestor_cases() {
-        use std::path::Path;
-        // Sibling files share their parent.
-        assert_eq!(
-            common_ancestor(Path::new("/m/Album/01.flac"), Path::new("/m/Album/02.flac")),
-            Path::new("/m/Album")
-        );
-        // `a` is already an ancestor of `b`: keep `a`.
-        assert_eq!(
-            common_ancestor(Path::new("/m/Album"), Path::new("/m/Album/Disc1/01.flac")),
-            Path::new("/m/Album")
-        );
-        // Component-wise, not string-prefix: Album vs Album2 don't share /m/Album.
-        assert_eq!(
-            common_ancestor(Path::new("/m/Album/x"), Path::new("/m/Album2/y")),
-            Path::new("/m")
-        );
-        // Disjoint trees collapse to the root.
-        assert_eq!(
-            common_ancestor(Path::new("/a/b"), Path::new("/c/d")),
-            Path::new("/")
-        );
-    }
-
-    /// image_cover_priority decides which folder image wins as the cover when
-    /// the user makes no explicit pick: a name containing "cover" or "front"
-    /// (case-insensitive, anywhere in the name) ranks first, everything else
-    /// second. The fallback sort relies on this ordering.
-    #[test]
-    fn image_cover_priority_ranks_front_and_cover_first() {
-        assert_eq!(ImportService::image_cover_priority("Cover.jpg"), 0);
-        assert_eq!(ImportService::image_cover_priority("front.png"), 0);
-        assert_eq!(ImportService::image_cover_priority("FRONT.JPG"), 0);
-        assert_eq!(
-            ImportService::image_cover_priority("album-front-scan.jpg"),
-            0
-        );
-        assert_eq!(ImportService::image_cover_priority("Back.jpg"), 1);
-        assert_eq!(ImportService::image_cover_priority("inlay.png"), 1);
-        assert_eq!(ImportService::image_cover_priority("disc1.jpg"), 1);
-    }
-
-    /// Deterministic clock for the `apply_user_edit_to_seed` tests — the
-    /// exact instant is immaterial to what they assert (artist-row
-    /// preservation / rebuild), only that the same one feeds every row.
-    fn test_clock() -> FixedClock {
-        FixedClock(
-            chrono::DateTime::parse_from_rfc3339("2024-01-01T00:00:00Z")
-                .unwrap()
-                .with_timezone(&chrono::Utc),
-        )
-    }
-
-    // ── apply_identity_choice ──────────────────────────────────────────
-
-    fn mb_id_exact(group: &str, release: &str) -> crate::import::ReleaseIdentity {
-        crate::import::ReleaseIdentity {
-            source: crate::import::MetadataSource::MusicBrainz,
-            source_group_id: group.to_string(),
-            source_release_id: Some(release.to_string()),
-        }
-    }
-
-    fn discogs_id_exact(group: &str, release: &str) -> crate::import::ReleaseIdentity {
-        crate::import::ReleaseIdentity {
-            source: crate::import::MetadataSource::Discogs,
-            source_group_id: group.to_string(),
-            source_release_id: Some(release.to_string()),
-        }
-    }
-
-    fn mb_release_ref() -> crate::import::MetadataRef {
-        crate::import::MetadataRef::new("rel-mb", crate::import::MetadataSource::MusicBrainz)
-    }
-
-    #[test]
-    fn exact_choice_passes_mapper_output_through() {
-        let mapper_output = vec![
-            mb_id_exact("rg-mb", "rel-mb"),
-            discogs_id_exact("master-d", "rel-d"),
-        ];
-        let result = apply_identity_choice(
-            &mapper_output,
-            &crate::import::IdentityChoice::Exact {
-                release_ref: mb_release_ref(),
-            },
-        );
-        assert_eq!(result, mapper_output);
-    }
-
-    #[test]
-    fn approximate_choice_nulls_release_ids_on_every_row() {
-        // Both the primary identity row AND any cross-source row from
-        // url-rels mirror the user's choice — Approximate means a
-        // group-level claim across the board.
-        let mapper_output = vec![
-            mb_id_exact("rg-mb", "rel-mb"),
-            discogs_id_exact("master-d", "rel-d"),
-        ];
-        let result = apply_identity_choice(
-            &mapper_output,
-            &crate::import::IdentityChoice::Approximate {
-                release_ref: mb_release_ref(),
-            },
-        );
-        assert_eq!(result.len(), 2);
-        for id in &result {
-            assert!(
-                id.source_release_id.is_none(),
-                "Approximate must NULL source_release_id, got {id:?}"
-            );
-        }
-        // Group IDs survive — the claim is at the group level.
-        assert_eq!(result[0].source_group_id, "rg-mb");
-        assert_eq!(result[1].source_group_id, "master-d");
-    }
-
-    #[test]
-    fn unknown_choice_passes_empty_mapper_output_through() {
-        // The file-tag mapper emits an empty identity vec; the choice
-        // post-process is a no-op. Confirms Unknown writes zero
-        // `release_identities` rows even when paired with a mapper
-        // that somehow surfaces rows (defensive — file_tag_mapper
-        // never does, but the projection is the algebraic identity).
-        let result = apply_identity_choice(&[], &crate::import::IdentityChoice::Unknown);
-        assert!(result.is_empty());
-    }
-
-    // ── apply_user_edit_to_seed ────────────────────────────────────────
-
-    fn make_seed_album_release_track() -> (
-        crate::db::DbAlbum,
-        crate::db::DbRelease,
-        crate::db::DbTrack,
-        crate::db::DbArtist,
-    ) {
-        let now = chrono::Utc::now();
-        let artist = crate::db::DbArtist {
-            id: "artist-orig".to_string(),
-            name: "Artist Name".to_string(),
-            sort_name: Some("Artist Name".to_string()),
-            discogs_artist_id: None,
-            musicbrainz_artist_id: None,
-            created_at: now,
-        };
-        let album = crate::db::DbAlbum {
-            id: "album-1".to_string(),
-            title: "Album Title".to_string(),
-            artist_id: artist.id.clone(),
-            year: Some(2020),
-            primary_release_id: None,
-            is_compilation: false,
-            created_at: now,
-        };
-        let release = crate::db::DbRelease {
-            id: "release-1".to_string(),
-            album_id: album.id.clone(),
-            release_name: None,
-            pressing: crate::db::Pressing {
-                year: Some(2020),
-                format: Some("CD".to_string()),
-                label: Some("Label Name".to_string()),
-                catalog_number: Some("CAT-001".to_string()),
-                country: None,
-                barcode: None,
-            },
-            disc_id: None,
-            metadata_source: crate::db::ReleaseMetadataSource::MusicBrainz,
-            metadata_source_release_id: Some("rel-mb".to_string()),
-            remote: true,
-            source_folder_name: None,
-            content_hash: None,
-            album_loudness_lufs: None,
-            album_peak_linear: None,
-            created_at: now,
-        };
-        let track = crate::db::DbTrack {
-            id: "track-1".to_string(),
-            release_id: release.id.clone(),
-            title: "Original Title".to_string(),
-            side: 1,
-            track_number: Some(1),
-            duration_ms: Some(180000),
-            discogs_position: None,
-            created_at: now,
-        };
-        (album, release, track, artist)
-    }
-
-    #[test]
-    fn user_edit_overrides_seeded_pressing_fields() {
-        let (mut album, mut release, track, seed_artist) = make_seed_album_release_track();
-        let mut tracks = vec![track];
-        let mut artists = vec![seed_artist];
-        let mut album_artists = Vec::new();
-        let mut track_artists = Vec::new();
-
-        let edit = crate::import::ReleaseUserEdit {
-            album_title: "Edited Title".to_string(),
-            album_artist_names: vec!["Edited Artist".to_string()],
-            pressing: crate::import::PressingEdit {
-                year: Some(1995),
-                format: Some("Vinyl".to_string()),
-                label: Some("Edited Label".to_string()),
-                catalog_number: Some("EDIT-1".to_string()),
-                country: Some("JP".to_string()),
-                barcode: Some("4943674000000".to_string()),
-            },
-            tracks: vec![crate::import::TrackUserEdit {
-                title: "Edited Track".to_string(),
-                side: 1,
-                track_number: Some(1),
-                artist_names: vec![],
-            }],
-        };
-
-        apply_user_edit_to_seed(
-            &edit,
-            &mut album,
-            &mut release,
-            &mut tracks,
-            &mut artists,
-            &mut album_artists,
-            &mut track_artists,
-            &test_clock(),
-            &SequentialIdProvider::new("seed"),
-        )
-        .unwrap();
-
-        assert_eq!(album.title, "Edited Title");
-        assert_eq!(release.pressing.year, Some(1995));
-        assert_eq!(release.pressing.format.as_deref(), Some("Vinyl"));
-        assert_eq!(release.pressing.label.as_deref(), Some("Edited Label"));
-        assert_eq!(release.pressing.catalog_number.as_deref(), Some("EDIT-1"));
-        assert_eq!(release.pressing.country.as_deref(), Some("JP"));
-        assert_eq!(release.pressing.barcode.as_deref(), Some("4943674000000"));
-        assert_eq!(tracks[0].title, "Edited Track");
-
-        // The new album artist gets a placeholder DbArtist row so the
-        // import pipeline can canonicalize via find_or_create_artists.
-        assert!(artists.iter().any(|a| a.name == "Edited Artist"));
-        assert_eq!(
-            album.artist_id,
-            artists
-                .iter()
-                .find(|a| a.name == "Edited Artist")
-                .unwrap()
-                .id
-        );
-    }
-
-    #[test]
-    fn user_edit_can_fill_country_for_approximate_seed() {
-        // Approximate seed clears pressing fields; the user can supply
-        // them via the editor and the overlay applies the value.
-        let (mut album, mut release, track, seed_artist) = make_seed_album_release_track();
-        // Simulate the Approximate-cleared release row.
-        release.pressing = crate::db::Pressing::blank();
-        let mut tracks = vec![track];
-        let mut artists = vec![seed_artist];
-        let mut album_artists = Vec::new();
-        let mut track_artists = Vec::new();
-
-        let edit = crate::import::ReleaseUserEdit {
-            album_title: album.title.clone(),
-            album_artist_names: vec![artists[0].name.clone()],
-            pressing: crate::import::PressingEdit {
-                country: Some("JP".to_string()),
-                ..crate::import::PressingEdit::blank()
-            },
-            tracks: vec![crate::import::TrackUserEdit {
-                title: tracks[0].title.clone(),
-                side: tracks[0].side,
-                track_number: tracks[0].track_number,
-                artist_names: vec![],
-            }],
-        };
-
-        apply_user_edit_to_seed(
-            &edit,
-            &mut album,
-            &mut release,
-            &mut tracks,
-            &mut artists,
-            &mut album_artists,
-            &mut track_artists,
-            &test_clock(),
-            &SequentialIdProvider::new("seed"),
-        )
-        .unwrap();
-
-        assert_eq!(release.pressing.country.as_deref(), Some("JP"));
-        assert!(release.pressing.year.is_none());
-        assert!(release.pressing.format.is_none());
-    }
-
-    #[test]
-    fn user_edit_track_count_mismatch_is_an_error() {
-        let (mut album, mut release, track, seed_artist) = make_seed_album_release_track();
-        let mut tracks = vec![track];
-        let mut artists = vec![seed_artist];
-        let mut album_artists = Vec::new();
-        let mut track_artists = Vec::new();
-
-        let edit = crate::import::ReleaseUserEdit {
-            album_title: "T".to_string(),
-            album_artist_names: vec!["A".to_string()],
-            pressing: crate::import::PressingEdit::blank(),
-            // Two edits but seed has one track.
-            tracks: vec![
-                crate::import::TrackUserEdit {
-                    title: "X".to_string(),
-                    side: 1,
-                    track_number: Some(1),
-                    artist_names: vec![],
-                },
-                crate::import::TrackUserEdit {
-                    title: "Y".to_string(),
-                    side: 1,
-                    track_number: Some(2),
-                    artist_names: vec![],
-                },
-            ],
-        };
-
-        let err = apply_user_edit_to_seed(
-            &edit,
-            &mut album,
-            &mut release,
-            &mut tracks,
-            &mut artists,
-            &mut album_artists,
-            &mut track_artists,
-            &test_clock(),
-            &SequentialIdProvider::new("seed"),
-        )
-        .unwrap_err();
-        assert!(err.contains("Track count mismatch"), "got: {err}");
-    }
-
-    /// Source-id linkage on artist rows (e.g. `musicbrainz_artist_id`)
-    /// must survive a user edit that doesn't touch artist names. The
-    /// editor round-trips an unchanged artist field as the same string
-    /// it was seeded with, so the apply step must compare and short-
-    /// circuit rather than rebuild rows from name-only placeholders.
-    #[test]
-    fn user_edit_preserves_source_id_artist_rows_when_names_unchanged() {
-        let now = chrono::Utc::now();
-        // Seeded artist row carrying the MB id the mapper attached.
-        let seed_artist = crate::db::DbArtist {
-            id: "artist-mb".to_string(),
-            name: "Artist Name".to_string(),
-            sort_name: Some("Artist Name".to_string()),
-            discogs_artist_id: None,
-            musicbrainz_artist_id: Some("mb-artist-1".to_string()),
-            created_at: now,
-        };
-        let album = crate::db::DbAlbum {
-            id: "album-1".to_string(),
-            title: "Album Title".to_string(),
-            artist_id: seed_artist.id.clone(),
-            year: Some(2020),
-            primary_release_id: None,
-            is_compilation: false,
-            created_at: now,
-        };
-        let release = crate::db::DbRelease {
-            id: "release-1".to_string(),
-            album_id: album.id.clone(),
-            release_name: None,
-            pressing: crate::db::Pressing {
-                year: Some(2020),
-                format: None,
-                label: None,
-                catalog_number: None,
-                country: None,
-                barcode: None,
-            },
-            disc_id: None,
-            metadata_source: crate::db::ReleaseMetadataSource::MusicBrainz,
-            metadata_source_release_id: Some("rel-mb".to_string()),
-            remote: true,
-            source_folder_name: None,
-            content_hash: None,
-            album_loudness_lufs: None,
-            album_peak_linear: None,
-            created_at: now,
-        };
-        let track = crate::db::DbTrack {
-            id: "track-1".to_string(),
-            release_id: release.id.clone(),
-            title: "Track Title".to_string(),
-            side: 1,
-            track_number: Some(1),
-            duration_ms: None,
-            discogs_position: None,
-            created_at: now,
-        };
-        // Seeded track credit pointing at the MB-id-bearing artist.
-        let seed_track_artist = crate::db::DbTrackArtist::new(
-            &track.id,
-            &seed_artist.id,
-            0,
-            "track-artist-1".to_string(),
-            now,
-        );
-
-        let mut album = album;
-        let mut release = release;
-        let mut tracks = vec![track];
-        let mut artists = vec![seed_artist.clone()];
-        let mut album_artists = Vec::<crate::db::DbAlbumArtist>::new();
-        let mut track_artists = vec![seed_track_artist.clone()];
-
-        // The user changes pressing fields but leaves artist names
-        // alone. The track's edit ships `artist_names = []` because
-        // the editor's "no override" form maps to empty when the
-        // track's credit equals the album's.
-        let edit = crate::import::ReleaseUserEdit {
-            album_title: album.title.clone(),
-            album_artist_names: vec![seed_artist.name.clone()],
-            pressing: crate::import::PressingEdit {
-                year: Some(1995),
-                ..crate::import::PressingEdit::blank()
-            },
-            tracks: vec![crate::import::TrackUserEdit {
-                title: tracks[0].title.clone(),
-                side: tracks[0].side,
-                track_number: tracks[0].track_number,
-                artist_names: vec![],
-            }],
-        };
-
-        apply_user_edit_to_seed(
-            &edit,
-            &mut album,
-            &mut release,
-            &mut tracks,
-            &mut artists,
-            &mut album_artists,
-            &mut track_artists,
-            &test_clock(),
-            &SequentialIdProvider::new("seed"),
-        )
-        .unwrap();
-
-        // The MB-id-bearing artist row must still exist with its
-        // source binding intact — no fresh placeholder created.
-        assert_eq!(artists.len(), 1, "no extra placeholder rows expected");
-        assert_eq!(
-            artists[0].musicbrainz_artist_id.as_deref(),
-            Some("mb-artist-1"),
-            "MB artist id must survive the edit",
-        );
-        assert_eq!(
-            album.artist_id, seed_artist.id,
-            "album.artist_id should still reference the seeded row",
-        );
-
-        // Track credit must still reference the seeded artist row.
-        assert_eq!(track_artists.len(), 1);
-        assert_eq!(track_artists[0].artist_id, seed_artist.id);
-    }
-
-    /// User-renaming an artist must rebuild the credit rows. The new
-    /// name has no source binding, so the inserted `DbArtist` row
-    /// carries `None` for both source ids.
-    #[test]
-    fn user_edit_renaming_album_artist_rebuilds_credits() {
-        let (mut album, mut release, track, seed_artist) = make_seed_album_release_track();
-        let mut tracks = vec![track];
-        let mut artists = vec![seed_artist.clone()];
-        let mut album_artists = Vec::new();
-        let mut track_artists = Vec::new();
-
-        let edit = crate::import::ReleaseUserEdit {
-            album_title: album.title.clone(),
-            album_artist_names: vec!["Different Artist".to_string()],
-            pressing: crate::import::PressingEdit::blank(),
-            tracks: vec![crate::import::TrackUserEdit {
-                title: tracks[0].title.clone(),
-                side: tracks[0].side,
-                track_number: tracks[0].track_number,
-                artist_names: vec![],
-            }],
-        };
-
-        apply_user_edit_to_seed(
-            &edit,
-            &mut album,
-            &mut release,
-            &mut tracks,
-            &mut artists,
-            &mut album_artists,
-            &mut track_artists,
-            &test_clock(),
-            &SequentialIdProvider::new("seed"),
-        )
-        .unwrap();
-
-        let new_artist = artists
-            .iter()
-            .find(|a| a.name == "Different Artist")
-            .expect("new placeholder should be inserted");
-        assert!(new_artist.musicbrainz_artist_id.is_none());
-        assert!(new_artist.discogs_artist_id.is_none());
-        assert_eq!(album.artist_id, new_artist.id);
-    }
-}
+mod tests;
